Admin and Student control class updates

+++ controller/StudentControl.class.php
@@ -1,16 +1,20 @@
 <?php
 
+// Import the Connection class and the Student model files
 require_once('./controller/database/Connection.class.php');
 require_once('./model/Student.class.php');
 
+// Global variable
+$connection = new Connection('./controller/database/config.ini');
+
 class StudentControl {
 
+	// Create a new student register in database
 	public function create($obj) {
 		try {
-			if (get_class($obj) == "Student") {
-				$connection = new Connection('./controller/database/config.ini');
-				$query = "INSERT INTO student VALUES (:name, :email, :gender, :course);";
-				$command = $connection->getPDO()->prepare($query);
+			if (get_class($obj) === "Student") {
+				global $connection;
+				$command = $connection->getPDO()->prepare("INSERT INTO student VALUES (:name, :email, :gender, :course);");
 				
 				$name = $obj->getName();
 				$email = $obj->getEmail();
@@ -35,9 +39,10 @@ class StudentControl {
 		}
 	}
 
+	// Select and return all student registers from database
 	public function read() {
 		try {
-			$connection = new Connection('./controller/database/config.ini');
+			global $connection;
 			$command = $connection->getPDO()->prepare("SELECT * FROM student");
 			if ($command->execute()) {
 				$data = $command->fetchAll(PDO::FETCH_CLASS, "Student");
@@ -52,9 +57,10 @@ class StudentControl {
 		}
 	}
 
+	// Return only one student register from database
 	public function readOne($email) {
 		try {
-			$connection = new Connection('./controller/database/config.ini');
+			global $connection;
 			$command = $connection->getPDO()->prepare("SELECT * FROM student WHERE email = :email;");
 			$command->bindParam("email", $email);
 			if ($command->execute()) {
@@ -70,47 +76,53 @@ class StudentControl {
 		}
 	}
 
+	// Update student register on database
 	public function update($obj) {
 		try {
-			$connection = new Connection('./controller/database/config.ini');
-			$command = $connection->getPDO()->prepare("UPDATE student SET name = :name, gender = :gender, course = :course WHERE email = :email");
+			if (get_class($obj) === "Student") {
+				global $connection;
+				$command = $connection->getPDO()->prepare("UPDATE student SET name = :name, gender = :gender, course = :course WHERE email = :email");
 
-			$name = $obj->getName();
-			$gender = $obj->getGender();
-			$course = $obj->getCourse();
-			$email = $obj->getEmail();
+				$name = $obj->getName();
+				$gender = $obj->getGender();
+				$course = $obj->getCourse();
+				$email = $obj->getEmail();
 
-			$command->bindParam("name", $name);
-			$command->bindParam("gender", $gender);
-			$command->bindParam("course", $course);
-			$command->bindParam("email", $email);
+				$command->bindParam("name", $name);
+				$command->bindParam("gender", $gender);
+				$command->bindParam("course", $course);
+				$command->bindParam("email", $email);
 
-			if ($command->execute()) {
-				$connection->closeConnection();
-				return true;
-			} else {
-				$connection->closeConnection();
-				return false;
+				if ($command->execute()) {
+					$connection->closeConnection();
+					return true;
+				} else {
+					$connection->closeConnection();
+					return false;
+				}
 			}
 		} catch (PDOException $e) {
 			echo "Error: {$e->getMessage()}";
 		}
 	}
 
+	// Delete one student register from database
 	public function delete($obj) {
 		try {
-			$connection = new Connection('./controller/database/config.ini');
-			$command = $connection->getPDO()->prepare("DELETE FROM student WHERE email = :email;");
+			if (get_class($obj) === "Student") {
+				global $connection;
+				$command = $connection->getPDO()->prepare("DELETE FROM student WHERE email = :email;");
 
-			$email = $obj->getEmail();
-			$command->bindParam("email", $email);
+				$email = $obj->getEmail();
+				$command->bindParam("email", $email);
 
-			if ($command->execute()) {
-				$connection->closeConnection();
-				return true;
-			} else {
-				$connection->closeConnection();
-				return false;
+				if ($command->execute()) {
+					$connection->closeConnection();
+					return true;
+				} else {
+					$connection->closeConnection();
+					return false;
+				}
 			}
 		} catch (PDOException $e) {
 			echo "Error: {$e->getMessage()}";
